* More code refinement for Bits++ GPU implementation. Shader fine-tuning for   better accuracy.
* Improved correctness tests.

* More bugs fixed in the old Bits++ toolbox.

-> Works perfectly well on MBP X1600. When fed with integral intensity levels,
   results of Bits++ TB and GPU TB match. When fed with random numbers, the
   GPU implementation is 1 unit off worst case, due to roundoff errors and
   single precision floats instead of double precision floats. Off by one
   should be more than acceptable in real world.

=> Test on NVidia pending.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@695 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ Psychtoolbox/PsychHardware/BitsPlusToolbox/BitsPlusPackMonoImage.m
@@ -4,7 +4,7 @@ function theColorImage = BitsPlusPackMonoImage(theMonoImage)
 % In Mono++ mode, the Bits++ box uses the red and green
 % channels to provide 14-bits per pixel true intensity
 % resolution.  The blue channel is set to 0 to let the
-% monochromatic image through.   Emprically, we concluded
+% monochromatic image through.   Empirically, we concluded
 % that the packing is left adjusted.  That is, the 8 MSB
 % of the 14-bit input go into the red channel, and the 
 % 6 LSB get left aligned in the green channel, with
@@ -15,9 +15,10 @@ function theColorImage = BitsPlusPackMonoImage(theMonoImage)
 % 11/17/03  dhb, ip	Wrote it.
 % 8/13/04	dhb		Fix bug, the data were not packed quite right.
 % 2/26/07   mk      Bugfix for LSB conversion: Added modulo operation.
+% 3/01/07   mk      Bugfix for MSB conversion: Added floor operation.
 
 [m,n] = size(theMonoImage);
-theColorImage = zeros(m,n,3);
+theColorImage = uint8(zeros(m,n,3));
 
 % Red channel gets high 8 bits
 theColorImage(:,:,1) = uint8(floor(theMonoImage/256));
